[DM-23507] Add a trivial test
This is just enough to let tests run so that the Travis CI status
can clear.  More extensive tests will be added later.

Add a new optional parameter to server to allow the test to not
start the user mapper, since it immediately tries to talk to Slack
and needs a valid Slack token.