Use structlog for mapper logging
Rather than initializing the root logger, change SlackGitHubMapper
to take an optional logger and get a logger from structlog to pass
in.  This will honor the logging settings for the application
without affecting other users of the root logger, such as the
slack module (which is very chatty at DEBUG).